Migration to v4.7

+++ tests/08-native-runs/mqtt-client.sh
@@ -23,10 +23,13 @@ CLIENT_TESTLOG=$TEST_NAME.testlog
 CLIENT_ERR=$TEST_NAME.err
 MOSQ_SUB_LOG=mosquitto_sub.log
 MOSQ_SUB_ERR=mosquitto_sub.err
+PUB_PAYLD=1
 
 # Start mosquitto server
 echo "Starting mosquitto daemon"
-mosquitto &> /dev/null &
+echo "listener 1883" > mosquitto.conf
+echo "allow_anonymous true" >> mosquitto.conf
+mosquitto -c mosquitto.conf &> /dev/null &
 MOSQID=$!
 sleep 2
 
@@ -38,6 +41,7 @@ sleep 2
 
 # Starting Contiki-NG native node
 echo "Starting native node"
+make -C $CODE_DIR -B TARGET=native clean
 make -C $CODE_DIR -B TARGET=native \
   DEFINES=MQTT_CLIENT_CONF_ORG_ID=\\\"travis-test\\\",MQTT_CLIENT_CONF_LOG_LEVEL=LOG_LEVEL_DBG,MQTT_CONF_VERSION=MQTT_PROTOCOL_VERSION_$MQTT_VERSION \
   > make.log 2> make.err
@@ -48,7 +52,8 @@ CPID=$!
 sleep 45
 
 # Send a publish to the mqtt client
-mosquitto_pub -m "1" -t iot-2/cmd/leds/fmt/json
+echo "Publishing"
+mosquitto_pub -m "$PUB_PAYLD" -t iot-2/cmd/leds/fmt/json
 
 echo "Closing native node"
 sleep 2
@@ -65,10 +70,13 @@ kill_bg $MSUBID
 # Success criteria:
 # * mosquitto_sub output not empty
 # * mqtt-client.native output contains "MQTT SUB"
+# * mqtt-client.native payload output matches published payload
 # * Valgrind off or 0 errors reported
 SUB_RCV=`grep "MQTT SUB" $CLIENT_LOG`
+SUB_RCV_PAYLD=`sed -rn "s/.*chunk='(.*)'/\1/p" $CLIENT_LOG`
 VALGRIND_NO_ERR=`grep "ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 0 from 0)" $CLIENT_ERR`
-if [ -s "$MOSQ_SUB_LOG" -a -n "$SUB_RCV" ] && [ -z "$VALGRIND_CMD" -o -n "$VALGRIND_NO_ERR" ]
+
+if [ -s "$MOSQ_SUB_LOG" -a -n "$SUB_RCV" -a "$SUB_RCV_PAYLD" -eq "$PUB_PAYLD" ] && [ -z "$VALGRIND_CMD" -o -n "$VALGRIND_NO_ERR" ]
 then
   cp $CLIENT_LOG $CLIENT_TESTLOG
   printf "%-32s TEST OK\n" "$TEST_NAME" | tee $CLIENT_TESTLOG;
@@ -87,6 +95,7 @@ rm make.log
 rm make.err
 rm $CLIENT_LOG $CLIENT_ERR
 rm $MOSQ_SUB_LOG $MOSQ_SUB_ERR
+rm mosquitto.conf
 
 # We do not want Make to stop -> Return 0
 # The Makefile will check if a log contains FAIL at the end
